refactor(web): ServiceAccountsController to Java
